Working board tiles read/write.

+++ fileIO/ewritestream.h
@@ -34,7 +34,7 @@ public:
 
     template <typename T>
     inline eWriteStream& operator<<(const T val) {
-        const int32_t val32_t = val;
+        const int32_t val32_t = static_cast<int32_t>(val);
         write(&val32_t, sizeof(int32_t));
         return *this;
     }
